Generate skeleton for sprite interactions
Generate a random Charmander sprite somewhere on the grid, and a Squirtle
sprite at a fixed location on the grid. The player controls Charmander
with the arrow keys. Motion is confined to a grid with square size 24x24.
Collisions with Squirtle result in teleportation to a random location on
the grid.