Rewrite level rendering
Rewrite level rendering to be done through a LEVEL struct, rather than
through another SPRITE. This allows me to add more information about the
levels in the future, like where is walkable, where items and enemies are,
and more.

A good deal of superfluous code was also removed (item and enemy
generation), since they're not needed yet. They were originally shoehorned
in for the purpose of hitting the checkpoint for the assignment.