Refactor sprite to have position relative to level
It makes a lot more sense for the player and enemies to have their
sprites' positions stored relative to the level, rather than the screen,
given that all of the gameplay logic is done relative to the level, and
ONLY rendering uses the screen. This created RenderSpriteOnLevel to render
a sprite relative to its level and updated the code.

As a bonus, the enemies now stay in their location on the map without my
having to change anything. This means we don't have to update the
positions of literally everything when we move the player (that is, move
the map relative to the screen).