Make enemies move towards player when close enough
Enemy sprites now move towards the player when they are within a certain
distance of the player. The pathfinding is very primitive (figure out
direction of greatest change and try to move in it), so enemies frequently
get stuck behind walls or at corners, but writing more complicated
pathfinding in assembly is not worth it.

Doing this required the creation of GetBearing to determine the direction,
SpriteDistance to determine the direction between two sprites, and AbsVal
as a helper.