Add music; fix rendering bugs
Added music to the game -- this takes the form of a music note tile on the
level that the player can choose to step on, which starts music playing.

Fixed some rendering bugs where things were getting rendered in the wrong
order (e.g., treasure boxes on top of enemies).