Detect collisions between enemy attack and player
This adds a lot of new code to detect collisions between enemy attacks and
the player. This was made more complicated than it probably needed to be
because of design decisions made early on, when I wasn't thinking about
how to handle enemy attacks.

In particular, we add various fields to the SPRITE structure that contain
information about the attack associated with that SPRITE. These fields are
used for collision detection throughout; it will make sense to use these
for player attacks in the future.