Add berries and redesign item generation
Redesign and refactor item generation into a new function, GenerateItem,
which allows us to escape the kluge of using GenerateEnemy and overwriting
the inconvenient results. This function can be used for treasures, the
music note, and the newly added berries.

Add berries -- generate them randomly around the dungeon, and have them
restore the player's health (or increase max health if it is already
full). Regenerate occasionally.