Rewrite enemy attack handling
Completely rewrite enemy attack handling so that it all takes place in one
procedure, which takes care of checking the attack is active, checking
collisions with walls and the player, making new attacks when the enemy is
in range, and all other logic related to them.

This fixed the bug referenced in the previous commit message.

Also, implement minor UI tweaks to the right panel.